+++ CHANGELOG.md
@@ -49,12 +49,12 @@
 - `Arg.setNameAsLongName` is removed.
 - `Arg.setAllowedValues` is renamed to `Arg.setValidValues`.
 - `Arg.setSetting` is renamed to `Arg.setProperty`.
-- `Arg.unsetSetting` is renamed to `Arg.removeProperty`.
+- `Arg.unsetSetting` is renamed to `Arg.unsetProperty`.
 - `Arg.isSettingSet` is renamed to `Arg.hasProperty`.
 - `Command.takesSingleValue` and `Command.takesNValues` are removed, use new `Arg.positional` instead.
 - `Command.countArgs` is renamed to `Command.countPositionalArgs`.
 - `Command.setSetting` is renamed to `Command.setProperty`.
-- `Command.unsetSetting` is renamed to `Command.removeProperty`.
+- `Command.unsetSetting` is renamed to `Command.unsetProperty`.
 - `Command.isSettingSet` is renamed to `Command.hasProperty`.
 
 ## What's New